Styler now returns strings rather than printing

+++ pkg/codec/http_output.go
@@ -24,7 +24,7 @@ func ParseListener(l net.Listener, d *state.DaemonData) {
 }
 
 // TODO; move to codec.tcp.go
-func ParseNetConn(c net.Conn, requestNo uint, d *state.RequestData) {
+func ParseNetConn(c net.Conn, requestNo uint64, d *state.RequestData) {
 	now := time.Now()
 	d.TransportConnTime = &now
 	d.TransportConnNo = requestNo
